etap 2 beggining
