+++ Etap2/Data/Ball.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
 
-namespace Logic
+namespace Data
 {
     public class Ball : INotifyPropertyChanged
     {
